ml-progress-bot/requirements.txt
# torch~=0.3.0
matplotlib~=1.5.3
numpy~=1.13.0
python-telegram-bot~=9.0.0
scikit-learn~=0.19.0
seaborn~=0.8.1